agent/pyproject.toml
[project]
name = "historical-battle-analyzer-agent"
version = "0.0.1"
description = "LangChain-based FastAPI backend for the Historical Battle Analyzer."
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "langchain",
  "langchain-community",
  "langchain-openai",
  "langchain-google-genai",
  "httpx",
  "python-dotenv",
  "pydantic>=2",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

